Move text renderer util classes to internal package
They are not API, so they should be in an internal package.
